+++ test/test.c
@@ -8,7 +8,7 @@ char* voidp_to_str(void* data, size_t data_size)
     char* data_str = (char*)data;
     char* message = malloc((data_size + 1) * sizeof(char));
 
-    for (int i = 0; i < data_size; i++) {
+    for (size_t i = 0; i < data_size; i++) {
         message[i] = data_str[i];
     }
 
@@ -16,7 +16,7 @@ char* voidp_to_str(void* data, size_t data_size)
     return message;
 }
 
-void server_on_recv(int client_id, void* data, size_t data_size, void* arg)
+void server_on_recv(size_t client_id, void* data, size_t data_size, void* arg)
 {
     char* message = voidp_to_str(data, data_size);
     printf("Received data from client #%d: %s (size %ld)\n", client_id, message, data_size);
@@ -24,12 +24,12 @@ void server_on_recv(int client_id, void* data, size_t data_size, void* arg)
     free(message);
 }
 
-void server_on_connect(int client_id, void* arg)
+void server_on_connect(size_t client_id, void* arg)
 {
     printf("Client #%d connected\n", client_id);
 }
 
-void server_on_disconnect(int client_id, void* arg)
+void server_on_disconnect(size_t client_id, void* arg)
 {
     printf("Client #%d disconnected\n", client_id);
 }
